+++ localflavor/br/forms.py
@@ -4,6 +4,7 @@
 from __future__ import unicode_literals
 
 import re
+import warnings
 
 from django.core.validators import EMPTY_VALUES
 from django.forms import ValidationError
@@ -11,7 +12,8 @@ from django.forms.fields import CharField, Field, RegexField, Select
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 
-from localflavor.generic.forms import DeprecatedPhoneNumberFormFieldMixin
+from localflavor.compat import EmptyValueCompatMixin
+from localflavor.deprecation import DeprecatedPhoneNumberFormFieldMixin, RemovedInLocalflavor20Warning
 
 from .br_states import STATE_CHOICES
 
@@ -98,12 +100,12 @@ def dv_maker(v):
     return 0
 
 
-# TODO deprecate function because it's name is not PEP8 compliant, issue #258
 def DV_maker(v):  # noqa
+    warnings.warn('DV_maker is deprecated. Please use dv_maker instead.', RemovedInLocalflavor20Warning)
     return dv_maker(v)
 
 
-class BRCPFField(CharField):
+class BRCPFField(EmptyValueCompatMixin, CharField):
     """
     A form field that validates a CPF number or a CPF string.
 
@@ -124,8 +126,8 @@ class BRCPFField(CharField):
     def clean(self, value):
         """Value can be either a string in the format XXX.XXX.XXX-XX or an 11-digit number."""
         value = super(BRCPFField, self).clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return self.empty_value
         orig_value = value[:]
         if not value.isdigit():
             cpf = cpf_digits_re.search(value)
@@ -153,7 +155,7 @@ class BRCPFField(CharField):
         return orig_value
 
 
-class BRCNPJField(CharField):
+class BRCNPJField(EmptyValueCompatMixin, CharField):
     """
     A form field that validates input as `Brazilian CNPJ`_.
 
@@ -183,8 +185,8 @@ class BRCNPJField(CharField):
     def clean(self, value):
         """Value can be either a string in the format XX.XXX.XXX/XXXX-XX or a group of 14 characters."""
         value = super(BRCNPJField, self).clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return self.empty_value
         orig_value = value[:]
         if not value.isdigit():
             cnpj = cnpj_digits_re.search(value)
@@ -213,7 +215,7 @@ def mod_97_base10(value):
     return 98 - ((value * 100 % 97) % 97)
 
 
-class BRProcessoField(CharField):
+class BRProcessoField(EmptyValueCompatMixin, CharField):
     """
     A form field that validates a Legal Process(Processo) number or a Legal Process string.
 
@@ -232,8 +234,8 @@ class BRProcessoField(CharField):
     def clean(self, value):
         """Value can be either a string in the format NNNNNNN-DD.AAAA.J.TR.OOOO or an 20-digit number."""
         value = super(BRProcessoField, self).clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return self.empty_value
 
         orig_value = value[:]
         if not value.isdigit():