+++ localflavor/sg/forms.py
@@ -3,14 +3,15 @@
 from __future__ import unicode_literals
 
 import re
+import warnings
 
-from django.core.validators import EMPTY_VALUES
 from django.forms import ValidationError
 from django.forms.fields import CharField, RegexField
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 
-from localflavor.generic.forms import DeprecatedPhoneNumberFormFieldMixin
+from localflavor.compat import EmptyValueCompatMixin
+from localflavor.deprecation import DeprecatedPhoneNumberFormFieldMixin, RemovedInLocalflavor20Warning
 
 PHONE_DIGITS_RE = re.compile(r'^[689](\d{7})$')
 
@@ -35,7 +36,7 @@ class SGPostCodeField(RegexField):
         super(SGPostCodeField, self).__init__(r'^\d{6}$', *args, **kwargs)
 
 
-class SGPhoneNumberField(CharField, DeprecatedPhoneNumberFormFieldMixin):
+class SGPhoneNumberField(EmptyValueCompatMixin, CharField, DeprecatedPhoneNumberFormFieldMixin):
     """
     A form field that validates input as a Singapore phone number.
 
@@ -51,8 +52,8 @@ class SGPhoneNumberField(CharField, DeprecatedPhoneNumberFormFieldMixin):
     def clean(self, value):
         """Validate a phone number. Strips parentheses, whitespace and hyphens."""
         super(SGPhoneNumberField, self).clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return self.empty_value
         value = re.sub('(\(|\)|\s+|-)', '', force_text(value))
         phone_match = PHONE_DIGITS_RE.search(value)
         if phone_match:
@@ -60,8 +61,7 @@ class SGPhoneNumberField(CharField, DeprecatedPhoneNumberFormFieldMixin):
         raise ValidationError(self.error_messages['invalid'])
 
 
-# TODO change to a pep8 compatible class name
-class SGNRIC_FINField(CharField):  # noqa
+class SGNRICFINField(EmptyValueCompatMixin, CharField):
     """
     A form field that validates input as a Singapore National Registration.
 
@@ -90,9 +90,9 @@ class SGNRIC_FINField(CharField):  # noqa
 
         Strips whitespace.
         """
-        super(SGNRIC_FINField, self).clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        super(SGNRICFINField, self).clean(value)
+        if value in self.empty_values:
+            return self.empty_value
         value = re.sub('(\s+)', '', force_text(value.upper()))
         match = NRIC_FIN_RE.search(value)
         if not match:
@@ -112,3 +112,10 @@ class SGNRIC_FINField(CharField):  # noqa
             return value
 
         raise ValidationError(self.error_messages['invalid'])
+
+
+class SGNRIC_FINField(SGNRICFINField):  # noqa
+    def __init__(self, *args, **kwargs):
+        warnings.warn('SGNRIC_FINField is deprecated. Please use SGNRICFINField instead.',
+                      RemovedInLocalflavor20Warning)
+        super(SGNRIC_FINField, self).__init__(*args, **kwargs)